Mana UI and Mana System
Built ManaBarUI and ManaSystem. When GainMana() is called, it can only be gained up until the maxMana level, when IncreaseMaxMana() is called both the currentMana and maxMana levels increase, and when a player tries to use an ability by calling ConsumeMana(), the manaCost is passed as a parameter to determine if the player has enough mana to cast the respective spell. If they have enough mana, true is returned and the manaCost is deducted from currentMana. If they don't have enough mana to cast a spell false is returned. The ManaBarUI script merely changes the fill amount of the manaFill canvas GameObject. This script is called whenever mana is consumed or gained in the ManaSystem script.