Wrote the script for the dialog system
Implemented the majority of the dialog system including the slow text reveal and ensuring the text doesn't shift

ToDo: I still need to allow the player to skip the slow reveal and allow the player to be able to dismiss the notification somehow.

Also the scrolling text functionality will not work with the way text is currently revealed at all as the text is expected to not move in order for the slow reveal to function.