Updated Unity and uncommented code
I'm sorry for commenting other people's code. It was the only Unity would allow me to use cinemachine. For some reason I get an error with "using EasyButtons;" and I don't know why. Is there a package I need to implement? Anyway here it is again. Same instructions as the last commit.