Updated Crumbling Floor Script
Now the platform triggers the warning sound when jumped on then begins the crumbling (destruction of the platform) when the player has been on the platform for 1 second.

Currently trying to fix the audio for when the floor begins to fall apart because at the moment it spamming the crumblingfloor audio clip while touching the platform during the crumbling phase.